+++ src/main/java/meshIneBits/gui/utilities/AboutDialogWindow.java
@@ -1,0 +1,99 @@
+/*
+ * MeshIneBits is a Java software to disintegrate a 3d mesh (model in .stl)
+ * into a network of standard parts (called "Bits").
+ *
+ * Copyright (C) 2016  Thibault Cassard & Nicolas Gouju.
+ * Copyright (C) 2017-2018  TRAN Quoc Nhat Han.
+ * Copyright (C) 2018 Vallon BENJAMIN.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package meshIneBits.gui.utilities;
+
+import meshIneBits.gui.IconLoader;
+import meshIneBits.util.Logger;
+
+import javax.swing.*;
+import java.awt.*;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.io.File;
+import java.io.IOException;
+import java.util.Objects;
+
+public class AboutDialogWindow extends JDialog {
+    private static final long serialVersionUID = -3389839563563221684L;
+
+    public AboutDialogWindow(JFrame parent, String title, boolean modal) {
+        super(parent, title, modal);
+
+        // Visual options
+        Image windowIcon = IconLoader.get("icon.png").getImage();
+        this.setIconImage(windowIcon);
+        this.setSize(350, 200);
+        this.setLocationRelativeTo(null);
+        this.setResizable(false);
+
+        // Setting up the dialog
+        JPanel jp = new JPanel();
+        jp.setLayout(new BoxLayout(jp, BoxLayout.PAGE_AXIS));
+
+        JLabel bg = new JLabel("");
+        ImageIcon icon = IconLoader.get("MeshIneBits.png", 248, 42);
+        bg.setIcon(icon);
+        bg.setFont(new Font(null, Font.BOLD | Font.ITALIC, 120));
+        bg.setForeground(new Color(0, 0, 0, 8));
+        bg.setAlignmentX(Component.CENTER_ALIGNMENT);
+
+        JLabel copyrightLabel1 = new JLabel("Copyright© 2016 Thibault Cassard & Nicolas Gouju.");
+        copyrightLabel1.setAlignmentX(Component.CENTER_ALIGNMENT);
+        JLabel copyrightLabel2 = new JLabel("Copyright© 2017-2018 TRAN Quoc Nhat Han.");
+        copyrightLabel2.setAlignmentX(Component.CENTER_ALIGNMENT);
+        JLabel copyrightLabel3 = new JLabel("Copyright© 2018 Vallon BENJAMIN.");
+        copyrightLabel3.setAlignmentX(Component.CENTER_ALIGNMENT);
+
+        JButton helpFileBtn = new JButton("Open help file (PDF format)");
+        helpFileBtn.setAlignmentX(Component.CENTER_ALIGNMENT);
+
+        jp.add(new JLabel(" "));
+        jp.add(bg);
+        jp.add(copyrightLabel1);
+        jp.add(copyrightLabel2);
+        jp.add(copyrightLabel3);
+        jp.add(new JLabel(" "));
+        jp.add(helpFileBtn);
+        this.getContentPane().add(jp, BorderLayout.CENTER);
+
+        // Actions listener
+        helpFileBtn.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                AboutDialogWindow.this.dispose();
+                Desktop dt = Desktop.getDesktop();
+                try {
+                    dt.open(new File(
+                            Objects.requireNonNull(this.getClass()
+                                    .getClassLoader()
+                                    .getResource("resources/help.pdf"))
+                                    .getPath()));
+                } catch (IOException e1) {
+                    Logger.error("Failed to load help file");
+                }
+            }
+        });
+
+        this.setVisible(true);
+    }
+}
